requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.5
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
requests==2.28.0
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.1
